# api/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "ChampionshipBetRecord": {
        "properties": {
          "bet_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bet Time"
          },
          "champion": {
            "title": "Champion",
            "type": "string"
          },
          "season": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              }
            ],
            "title": "Season"
          },
          "team": {
            "title": "Team",
            "type": "string"
          },
          "user_nome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Nome"
          },
          "vice": {
            "title": "Vice",
            "type": "string"
          }
        },
        "required": [
          "champion",
          "vice",
          "team",
          "season"
        ],
        "title": "ChampionshipBetRecord",
        "type": "object"
      },
      "ChampionshipBetRequest": {
        "additionalProperties": false,
        "properties": {
          "champion": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Champion",
            "type": "string"
          },
          "team": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Team",
            "type": "string"
          },
          "vice": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Vice",
            "type": "string"
          }
        },
        "required": [
          "champion",
          "vice",
          "team"
        ],
        "title": "ChampionshipBetRequest",
        "type": "object"
      },
      "ChampionshipResponse": {
        "properties": {
          "all_bets": {
            "items": {
              "$ref": "#/components/schemas/ChampionshipBetRecord"
            },
            "title": "All Bets",
            "type": "array"
          },
          "can_bet": {
            "title": "Can Bet",
            "type": "boolean"
          },
          "current_bet": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChampionshipBetRecord"
              },
              {
                "type": "null"
              }
            ]
          },
          "deadline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline"
          },
          "deadline_message": {
            "title": "Deadline Message",
            "type": "string"
          },
          "drivers": {
            "items": {
              "type": "string"
            },
            "title": "Drivers",
            "type": "array"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/ChampionshipBetRecord"
            },
            "title": "History",
            "type": "array"
          },
          "official_result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChampionshipResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "season": {
            "title": "Season",
            "type": "string"
          },
          "teams": {
            "items": {
              "type": "string"
            },
            "title": "Teams",
            "type": "array"
          }
        },
        "required": [
          "season",
          "drivers",
          "teams",
          "history",
          "all_bets",
          "can_bet",
          "deadline_message"
        ],
        "title": "ChampionshipResponse",
        "type": "object"
      },
      "ChampionshipResult": {
        "properties": {
          "champion": {
            "title": "Champion",
            "type": "string"
          },
          "season": {
            "title": "Season",
            "type": "string"
          },
          "team": {
            "title": "Team",
            "type": "string"
          },
          "vice": {
            "title": "Vice",
            "type": "string"
          }
        },
        "required": [
          "season",
          "champion",
          "vice",
          "team"
        ],
        "title": "ChampionshipResult",
        "type": "object"
      },
...
      "DriverRequest": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "number": {
            "default": 0,
            "maximum": 99.0,
            "minimum": 0.0,
            "title": "Number",
            "type": "integer"
          },
          "status": {
            "default": "Ativo",
            "maxLength": 30,
            "title": "Status",
            "type": "string"
          },
          "team": {
            "default": "",
            "maxLength": 120,
            "title": "Team",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "DriverRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RaceRequest": {
        "additionalProperties": false,
        "properties": {
          "circuit_id": {
            "anyOf": [
              {
                "maxLength": 80,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Circuit Id"
          },
          "date": {
            "maxLength": 30,
            "minLength": 8,
            "title": "Date",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "race_status": {
            "default": "Pendente",
            "maxLength": 30,
            "title": "Race Status",
            "type": "string"
          },
          "time": {
            "default": "",
            "maxLength": 20,
            "title": "Time",
            "type": "string"
          },
          "type": {
            "default": "Normal",
            "maxLength": 30,
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "name",
          "date"
        ],
        "title": "RaceRequest",
        "type": "object"
      },
...
      "ResultRequest": {
        "additionalProperties": false,
        "properties": {
          "positions": {
            "additionalProperties": true,
            "title": "Positions",
            "type": "object"
          },
          "retirements": {
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "title": "Retirements",
            "type": "array"
          }
        },
        "required": [
          "positions"
        ],
        "title": "ResultRequest",
        "type": "object"
      },
...
      "UserCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "password": {
            "maxLength": 1024,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          },
          "profile": {
            "pattern": "^(participante|admin|master|inativo)$",
            "title": "Profile",
            "type": "string"
          },
          "user_status": {
            "default": "ativo",
            "pattern": "^(ativo|inativo)$",
            "title": "User Status",
            "type": "string"
          }
        },
        "required": [
          "name",
          "email",
          "password",
          "profile"
        ],
        "title": "UserCreateRequest",
        "type": "object"
      },
      "UserResponse": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "must_change_password": {
            "default": false,
            "title": "Must Change Password",
            "type": "boolean"
          },
          "nome": {
            "title": "Nome",
            "type": "string"
          },
          "perfil": {
            "title": "Perfil",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "nome",
          "email",
          "perfil",
          "status"
        ],
        "title": "UserResponse",
        "type": "object"
      },
      "UserUpdateRequest": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "anyOf": [
              {
                "format": "email",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "must_change_password": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Must Change Password"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "profile": {
            "anyOf": [
              {
                "pattern": "^(participante|admin|master|inativo)$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile"
          },
          "user_status": {
            "anyOf": [
              {
                "pattern": "^(ativo|inativo)$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Status"
          }
        },
        "title": "UserUpdateRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "BF1 API",
    "version": "4.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin/drivers": {
      "post": {
        "operationId": "create_admin_driver_api_v1_admin_drivers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Admin Driver",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/drivers/{driver_id}": {
      "put": {
        "operationId": "update_admin_driver_api_v1_admin_drivers__driver_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "driver_id",
            "required": true,
            "schema": {
              "title": "Driver Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Admin Driver",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/races": {
      "post": {
        "operationId": "create_admin_race_api_v1_admin_races_post",
        "parameters": [
          {
            "in": "query",
            "name": "season",
            "required": true,
            "schema": {
              "pattern": "^\\d{4}$",
              "title": "Season",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Admin Race",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/races/{race_id}": {
      "put": {
        "operationId": "update_admin_race_api_v1_admin_races__race_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "race_id",
            "required": true,
            "schema": {
              "title": "Race Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "season",
            "required": true,
            "schema": {
              "pattern": "^\\d{4}$",
              "title": "Season",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Admin Race",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/races/{race_id}/result": {
      "put": {
        "operationId": "update_admin_result_api_v1_admin_races__race_id__result_put",
        "parameters": [
          {
            "in": "path",
            "name": "race_id",
            "required": true,
            "schema": {
              "title": "Race Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "season",
            "required": true,
            "schema": {
              "pattern": "^\\d{4}$",
              "title": "Season",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResultRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Admin Result",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/users": {
      "post": {
        "operationId": "create_admin_user_api_v1_admin_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Admin User",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/users/{user_id}": {
      "patch": {
        "operationId": "patch_admin_user_api_v1_admin_users__user_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Admin User",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/v1/championship": {
      "get": {
        "operationId": "championship_snapshot_api_v1_championship_get",
        "parameters": [
          {
            "in": "query",
            "name": "season",
            "required": true,
            "schema": {
              "maximum": 2100,
              "minimum": 2000,
              "title": "Season",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChampionshipResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Championship Snapshot",
        "tags": [
          "championship"
        ]
      }
    },
    "/api/v1/championship/bet": {
      "post": {
        "operationId": "championship_bet_api_v1_championship_bet_post",
        "parameters": [
          {
            "in": "query",
            "name": "season",
            "required": true,
            "schema": {
              "maximum": 2100,
              "minimum": 2000,
              "title": "Season",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChampionshipBetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChampionshipBetRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Championship Bet",
        "tags": [
          "championship"
        ]
      }
    },
    "/api/v1/championship/result": {
      "post": {
        "operationId": "championship_result_api_v1_championship_result_post",
        "parameters": [
          {
            "in": "query",
            "name": "season",
            "required": true,
            "schema": {
              "maximum": 2100,
              "minimum": 2000,
              "title": "Season",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChampionshipBetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChampionshipResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Championship Result",
        "tags": [
          "championship"
        ]
      }
    },
...
    "/api/v1/users/{user_id}": {
      "get": {
        "operationId": "get_user_api_v1_users__user_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get User",
        "tags": [
          "users"
        ]
      }
    }
  }
}